# Pre hash columns