--- a/gen/proto/go/frostdb/schema/v1alpha1/schema_vtproto.pb.go
+++ b/gen/proto/go/frostdb/schema/v1alpha1/schema_vtproto.pb.go
@@ -112,6 +112,16 @@ func (m *Column) MarshalToSizedBufferVT(dAtA []byte) (int, error) {
 		i -= len(m.unknownFields)
 		copy(dAtA[i:], m.unknownFields)
 	}
+	if m.Prehash {
+		i--
+		if m.Prehash {
+			dAtA[i] = 1
+		} else {
+			dAtA[i] = 0
+		}
+		i--
+		dAtA[i] = 0x20
+	}
 	if m.Dynamic {
 		i--
 		if m.Dynamic {
@@ -317,6 +327,9 @@ func (m *Column) SizeVT() (n int) {
 		n += 1 + l + sov(uint64(l))
 	}
 	if m.Dynamic {
+		n += 2
+	}
+	if m.Prehash {
 		n += 2
 	}
 	n += len(m.unknownFields)
@@ -642,6 +655,26 @@ func (m *Column) UnmarshalVT(dAtA []byte) error {
 				}
 			}
 			m.Dynamic = bool(v != 0)
+		case 4:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Prehash", wireType)
+			}
+			var v int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflow
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				v |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			m.Prehash = bool(v != 0)
 		default:
 			iNdEx = preIndex
 			skippy, err := skip(dAtA[iNdEx:])