--- a/proto/frostdb/schema/v1alpha1/schema.proto
+++ b/proto/frostdb/schema/v1alpha1/schema.proto
@@ -21,6 +21,7 @@ message Column {
     // Whether the column can dynamically expand.
     bool dynamic = 3;
     // Prehash the column before storing it. This is an optimization to speed up aggregation queries when this column is often aggregated. 
+    // This will create a separate non-dynamic column with the same name and the prefix "hashed." that contains the prehashed values.
     bool prehash = 4;
 }