This adds the ability to specify a column should be prehashed before inserted into the table. 

This allows to aggregator to skip performing the HashArray function for already prehashed columns. This comes with a small alloc cost for aggregations on non pre-hashed data.

```
name                  old time/op    new time/op    delta
Aggregation/sum-10       867µs ± 1%     874µs ± 2%    ~     (p=0.075 n=10+10)
Aggregation/count-10     869µs ± 1%     872µs ± 1%    ~     (p=0.218 n=10+10)
Aggregation/max-10       871µs ± 1%     876µs ± 1%  +0.62%  (p=0.003 n=10+9)

name                  old alloc/op   new alloc/op   delta
Aggregation/sum-10      1.16MB ± 0%    1.16MB ± 0%  +0.22%  (p=0.000 n=10+10)
Aggregation/count-10    1.16MB ± 0%    1.16MB ± 0%  +0.22%  (p=0.000 n=8+10)
Aggregation/max-10      1.16MB ± 0%    1.16MB ± 0%  +0.23%  (p=0.000 n=10+10)

name                  old allocs/op  new allocs/op  delta
Aggregation/sum-10         825 ± 0%       844 ± 0%  +2.30%  (p=0.000 n=9+10)
Aggregation/count-10       824 ± 0%       843 ± 0%  +2.31%  (p=0.000 n=10+9)
Aggregation/max-10         824 ± 0%       844 ± 0%  +2.37%  (p=0.000 n=10+9)
```