@thorfour 2023-09-07
Comparison on large aggregations 
```
name              old time/op    new time/op    delta
_PreAggregate-10     300ms ± 2%     289ms ± 4%  -3.62%  (p=0.000 n=10+10)

name              old alloc/op   new alloc/op   delta
_PreAggregate-10     603MB ± 0%     612MB ± 0%  +1.37%  (p=0.000 n=10+8)

name              old allocs/op  new allocs/op  delta
_PreAggregate-10     5.80M ± 0%     5.28M ± 0%  -9.02%  (p=0.000 n=10+10)
```