@thorfour 2023-09-07
Comparison when the data is prehashd
```
name                  old time/op    new time/op    delta
Aggregation/sum-10       379µs ± 0%     299µs ± 3%  -21.04%  (p=0.000 n=9+10)
Aggregation/count-10     374µs ± 1%     293µs ± 2%  -21.81%  (p=0.000 n=9+10)
Aggregation/max-10       384µs ± 1%     301µs ± 0%  -21.69%  (p=0.000 n=9+9)

name                  old alloc/op   new alloc/op   delta
Aggregation/sum-10       400kB ± 0%     401kB ± 0%   +0.06%  (p=0.000 n=9+9)
Aggregation/count-10     401kB ± 0%     401kB ± 0%   +0.06%  (p=0.000 n=10+10)
Aggregation/max-10       400kB ± 0%     401kB ± 0%   +0.06%  (p=0.000 n=10+10)

name                  old allocs/op  new allocs/op  delta
Aggregation/sum-10         686 ± 0%       686 ± 0%     ~     (all equal)
Aggregation/count-10       685 ± 0%       685 ± 0%     ~     (p=0.087 n=10+10)
Aggregation/max-10         686 ± 0%       686 ± 0%     ~     (all equal)
```