@thorfour 2023-09-12
Latest benchmark
```
thor@thors-MacBook-Pro frostdb % benchstat BenchmarkAggregation BenchmarkAggregation-prehashed
name                  old time/op    new time/op    delta
Aggregation/sum-10       406µs ± 1%     338µs ± 1%  -16.55%  (p=0.000 n=10+10)
Aggregation/count-10     400µs ± 1%     332µs ± 0%  -16.97%  (p=0.000 n=9+9)
Aggregation/max-10       409µs ± 0%     342µs ± 0%  -16.27%  (p=0.000 n=10+10)

name                  old alloc/op   new alloc/op   delta
Aggregation/sum-10       426kB ± 0%     433kB ± 0%   +1.68%  (p=0.000 n=10+7)
Aggregation/count-10     424kB ± 0%     432kB ± 0%   +1.82%  (p=0.000 n=10+10)
Aggregation/max-10       426kB ± 0%     434kB ± 0%   +1.70%  (p=0.000 n=8+10)

name                  old allocs/op  new allocs/op  delta
Aggregation/sum-10         788 ± 0%       853 ± 0%   +8.25%  (p=0.000 n=10+10)
Aggregation/count-10       787 ± 0%       853 ± 0%   +8.39%  (p=0.000 n=10+10)
Aggregation/max-10         788 ± 0%       853 ± 0%   +8.25%  (p=0.000 n=10+10)
```

Got a small bit worse due to some bug fixes